Adds type info to the list of service endpoints
AWS recently released s3 *interface* endpoints, in addition to the
prior s3 *gateway* endpoints. Unfortunately, the service name for
each is exactly the same, which makes it impossible in a single
attribute to determine which type the user wanted.

This patch works around that by using name and type attributes for
each service, so the user can specify exactly the type of endpoint
they want.